initi second

+++ src/main/java/com/practise/practice/repository/EmployeeRepository.java
@@ -3,7 +3,8 @@ package com.practise.practice.repository;
 import com.practise.practice.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
-
+public interface EmployeeRepository extends JpaRepository<Employee,Long> {
     Employee findByName(String name);
+
+//    Employee findName(String name);
 }
